Render selected way info on mouse hover
While this is very cool, there are real reasons to do this. In order to
find a sane path finding algorithm for walking, we have to understand
what our data looks like. One area that is not clear to me at the moment
is how the different way tags actually relate to areas that I can and
can not access. Without a way to tie the data back to the map, there is
no quick way to catch up here.

Find the closest way to the mouse, and start adding info about it to the
overlay. Color in the selected way as well to make it clear that we are
reading the correct info.

Implementation is a little hairy, see comments in commit for a better
idea of what is happening