Find closest point on selected way
In order to request a path plan, we need to be able to map cursor x/y to
a start/end location. Right now we have the selected way, but where on
the way are we?

Walk the selected way and find the closest point to the cursor. Render
it on the map for visibility/debugability