Native re-write
Pull as much code out of javascript and into rust as possible.
Re-implement the UI layer in egui on top of the new cross platform core
app. Trim down javascript to call into core wasm binary.

There are a few reasons to do this

* It's cool
* It allows for more flexibility in the future. If at some point I
  decide I do not want this to be a web app it's easy to switch to just
  slap a different UI on top
* It allows me to use tooling on native code to improve. E.g. using
  perf I can get line by line breakdown of the most expensive portions
  of the app
* I don't have to re-parse the input json blob both in javascript and
  rust. This saves a couple seconds during startup
* It allows me to prototype ideas more quickly without needing them to
  work in WASM land. E.g. if I find a cool GPU algorithm in CUDA, I can
  integrate that on the native side first to see if it would give me
  a speed increase before spending the effort on re-implementing it in
  OpenGL shaders

Implementation uses glow for all the rendering code. For the most part
this just works on both web and my desktop. I went with egui for the
frontend because I know they can handle me just doing my own OpenGl
stuff under their app. I briefly considered trying to re-use the
frontend code so I didn't have to write it twice (e.g. tauri or
integrating webkit or something), but this didn't look like it would be
as simple of a path forward

Some real changes came in during the re-write and I'll just list what
I'm aware of here instead of trying to break the changes out or fix them

* node to way lookup is now a neighbor lookup. This came out of some
  profiling that noticed that 30% of our time during path planning was
  spent on neighbor lookups
* Hardcoded colors for bike routes have been lost, I'm am not sure how I
  want this implemented yet. Depending on where performance
  investigation goes a tag based system where costs can be calculated on
  GPU instead of via lookup may make more sense
* Regex highlighting now supports many highlights. UI for this feature
  has not been implemented on the HTML side, but works on egui

This is overall a little faster, although not great. On my machine
planning a path from UBC to burnaby is in the 100-130ms range, vs the
130-160ms range that it used to be. While those numbers do not seem
impressive, they do feel better. I guess when we're towards 100ms we've
almost doubled our FPS